Extract same logic code for ShowSlaveStatusDatabaseDiscoveryType and MGRDatabaseDiscoveryType
