+++ shardingsphere-features/shardingsphere-db-discovery/shardingsphere-db-discovery-provider/shardingsphere-db-discovery-mysql/src/main/java/org/apache/shardingsphere/dbdiscovery/mysql/MGRDatabaseDiscoveryType.java
@@ -43,7 +43,7 @@ import java.util.Properties;
  * MGR database discovery type.
  */
 @Slf4j
-public final class MGRDatabaseDiscoveryType implements DatabaseDiscoveryType {
+public final class MGRDatabaseDiscoveryType extends AbstractDatabaseDiscoveryType implements DatabaseDiscoveryType {
     
     private static final String PLUGIN_STATUS = "SELECT * FROM information_schema.PLUGINS WHERE PLUGIN_NAME='group_replication'";
     
@@ -55,7 +55,8 @@ public final class MGRDatabaseDiscoveryType implements DatabaseDiscoveryType {
     
     private static final String MEMBER_LIST = "SELECT MEMBER_HOST, MEMBER_PORT, MEMBER_STATE FROM performance_schema.replication_group_members";
     
-    private String oldPrimaryDataSource;
+    private static final String PRIMARY_DATA_SOURCE = "SELECT MEMBER_HOST, MEMBER_PORT FROM performance_schema.replication_group_members WHERE MEMBER_ID = " +
+            "(SELECT VARIABLE_VALUE FROM performance_schema.global_status WHERE VARIABLE_NAME = 'group_replication_primary_member')";
     
     @Getter
     @Setter
@@ -141,58 +142,13 @@ public final class MGRDatabaseDiscoveryType implements DatabaseDiscoveryType {
     }
     
     @Override
-    public void updatePrimaryDataSource(final String schemaName, final Map<String, DataSource> dataSourceMap, final Collection<String> disabledDataSourceNames, final String groupName) {
-        Map<String, DataSource> activeDataSourceMap = new HashMap<>(dataSourceMap);
-        if (!disabledDataSourceNames.isEmpty()) {
-            activeDataSourceMap.entrySet().removeIf(each -> disabledDataSourceNames.contains(each.getKey()));
-        }
-        String newPrimaryDataSource = determinePrimaryDataSource(activeDataSourceMap);
-        if (newPrimaryDataSource.isEmpty()) {
-            return;
-        }
-        if (!newPrimaryDataSource.equals(oldPrimaryDataSource)) {
-            oldPrimaryDataSource = newPrimaryDataSource;
-            ShardingSphereEventBus.getInstance().post(new PrimaryDataSourceChangedEvent(schemaName, groupName, newPrimaryDataSource));
-        }
-    }
-    
-    private String determinePrimaryDataSource(final Map<String, DataSource> dataSourceMap) {
-        String primaryDataSourceURL = findPrimaryDataSourceURL(dataSourceMap);
-        return findPrimaryDataSourceName(primaryDataSourceURL, dataSourceMap);
-    }
-    
-    private String findPrimaryDataSourceURL(final Map<String, DataSource> dataSourceMap) {
-        String result = "";
-        String sql = "SELECT MEMBER_HOST, MEMBER_PORT FROM performance_schema.replication_group_members WHERE MEMBER_ID = "
-                + "(SELECT VARIABLE_VALUE FROM performance_schema.global_status WHERE VARIABLE_NAME = 'group_replication_primary_member')";
-        for (DataSource each : dataSourceMap.values()) {
-            try (Connection connection = each.getConnection();
-                 Statement statement = connection.createStatement();
-                 ResultSet resultSet = statement.executeQuery(sql)) {
-                if (resultSet.next()) {
-                    return String.format("%s:%s", resultSet.getString("MEMBER_HOST"), resultSet.getString("MEMBER_PORT"));
-                }
-            } catch (final SQLException ex) {
-                log.error("An exception occurred while find primary data source url", ex);
+    protected String getPrimaryDataSourceURL(final Statement statement) throws SQLException {
+        try (ResultSet resultSet = statement.executeQuery(PRIMARY_DATA_SOURCE)) {
+            if (resultSet.next()) {
+                return String.format("%s:%s", resultSet.getString("MEMBER_HOST"), resultSet.getString("MEMBER_PORT"));
             }
+            return "";
         }
-        return result;
-    }
-    
-    private String findPrimaryDataSourceName(final String primaryDataSourceURL, final Map<String, DataSource> dataSourceMap) {
-        String result = "";
-        for (Entry<String, DataSource> entry : dataSourceMap.entrySet()) {
-            String url;
-            try (Connection connection = entry.getValue().getConnection()) {
-                url = connection.getMetaData().getURL();
-                if (null != url && url.contains(primaryDataSourceURL)) {
-                    return entry.getKey();
-                }
-            } catch (final SQLException ex) {
-                log.error("An exception occurred while find primary data source name", ex);
-            }
-        }
-        return result;
     }
     
     @Override
@@ -212,7 +168,7 @@ public final class MGRDatabaseDiscoveryType implements DatabaseDiscoveryType {
     
     private List<String> findMemberDataSourceURLs(final Map<String, DataSource> activeDataSourceMap) {
         List<String> result = new LinkedList<>();
-        try (Connection connection = activeDataSourceMap.get(oldPrimaryDataSource).getConnection();
+        try (Connection connection = activeDataSourceMap.get(getOldPrimaryDataSource()).getConnection();
              Statement statement = connection.createStatement()) {
             ResultSet resultSet = statement.executeQuery(MEMBER_LIST);
             while (resultSet.next()) {
@@ -277,11 +233,6 @@ public final class MGRDatabaseDiscoveryType implements DatabaseDiscoveryType {
                 }
             }
         }
-    }
-    
-    @Override
-    public String getPrimaryDataSource() {
-        return oldPrimaryDataSource;
     }
     
     @Override
