+++ shardingsphere-features/shardingsphere-db-discovery/shardingsphere-db-discovery-provider/shardingsphere-db-discovery-mysql/src/main/java/org/apache/shardingsphere/dbdiscovery/mysql/ShowSlaveStatusDatabaseDiscoveryType.java
@@ -40,11 +40,9 @@ import java.util.Properties;
  * Show slave status database discovery type.
  */
 @Slf4j
-public final class ShowSlaveStatusDatabaseDiscoveryType implements DatabaseDiscoveryType {
+public final class ShowSlaveStatusDatabaseDiscoveryType extends AbstractDatabaseDiscoveryType {
     
     private static final String SHOW_SLAVE_STATUS = "SHOW SLAVE STATUS";
-    
-    private String oldPrimaryDataSource;
     
     @Getter
     @Setter
@@ -54,62 +52,7 @@ public final class ShowSlaveStatusDatabaseDiscoveryType implements DatabaseDisco
     public void checkDatabaseDiscoveryConfiguration(final String schemaName, final Map<String, DataSource> dataSourceMap) {
         //TODO Check master-slave mode
     }
-    
-    //TODO Consider merge MGR 's get primary datasource method
-    @Override
-    public void updatePrimaryDataSource(final String schemaName, final Map<String, DataSource> dataSourceMap, final Collection<String> disabledDataSourceNames, final String groupName) {
-        Map<String, DataSource> activeDataSourceMap = new HashMap<>(dataSourceMap);
-        if (!disabledDataSourceNames.isEmpty()) {
-            activeDataSourceMap.entrySet().removeIf(each -> disabledDataSourceNames.contains(each.getKey()));
-        }
-        String primaryDatasource = determinePrimaryDataSource(activeDataSourceMap);
-        if (primaryDatasource.isEmpty()) {
-            return;
-        }
-        if (!primaryDatasource.equals(oldPrimaryDataSource)) {
-            oldPrimaryDataSource = primaryDatasource;
-            ShardingSphereEventBus.getInstance().post(new PrimaryDataSourceChangedEvent(schemaName, groupName, primaryDatasource));
-        }
-    }
-    
-    //TODO Consider merge MGR 's get primary datasource method
-    private String determinePrimaryDataSource(final Map<String, DataSource> dataSourceMap) {
-        String primaryDataSourceURL = findPrimaryDataSourceURL(dataSourceMap);
-        return findPrimaryDataSourceName(primaryDataSourceURL, dataSourceMap);
-    }
-    
-    //TODO Consider merge MGR 's get primary datasource method
-    private String findPrimaryDataSourceURL(final Map<String, DataSource> dataSourceMap) {
-        for (Entry<String, DataSource> entry : dataSourceMap.entrySet()) {
-            try (Connection connection = entry.getValue().getConnection();
-                 Statement statement = connection.createStatement();
-                 ResultSet resultSet = statement.executeQuery(SHOW_SLAVE_STATUS)) {
-                if (resultSet.next()) {
-                    return String.format("%s:%s", resultSet.getString("Master_Host"), resultSet.getString("Master_Port"));
-                }
-            } catch (SQLException ex) {
-                log.error("An exception occurred while find primary data source url", ex);
-            }
-        }
-        return null;
-    }
-    
-    private String findPrimaryDataSourceName(final String primaryDataSourceURL, final Map<String, DataSource> dataSourceMap) {
-        String result = "";
-        for (Entry<String, DataSource> entry : dataSourceMap.entrySet()) {
-            String url;
-            try (Connection connection = entry.getValue().getConnection()) {
-                url = connection.getMetaData().getURL();
-                if (null != url && url.contains(primaryDataSourceURL)) {
-                    return entry.getKey();
-                }
-            } catch (final SQLException ex) {
-                log.error("An exception occurred while find primary data source name", ex);
-            }
-        }
-        return result;
-    }
-    
+
     @Override
     public void updateMemberState(final String schemaName, final Map<String, DataSource> dataSourceMap, final Collection<String> disabledDataSourceNames) {
         Map<String, DataSource> activeDataSourceMap = new HashMap<>(dataSourceMap);
@@ -117,11 +60,12 @@ public final class ShowSlaveStatusDatabaseDiscoveryType implements DatabaseDisco
             activeDataSourceMap.entrySet().removeIf(each -> disabledDataSourceNames.contains(each.getKey()));
         }
         for (Entry<String, DataSource> entry : activeDataSourceMap.entrySet()) {
-            if (oldPrimaryDataSource.equals(entry.getKey())) {
+            if (getOldPrimaryDataSource().equals(entry.getKey())) {
                 continue;
             }
             determineDatasourceState(schemaName, entry.getKey(), entry.getValue());
         }
+    
     }
     
     private void determineDatasourceState(final String schemaName, final String datasourceName, final DataSource dataSource) {
@@ -138,6 +82,16 @@ public final class ShowSlaveStatusDatabaseDiscoveryType implements DatabaseDisco
         }
     }
     
+    @Override
+    protected String getPrimaryDataSourceURL(final Statement statement) throws SQLException {
+        try (ResultSet resultSet = statement.executeQuery(SHOW_SLAVE_STATUS)) {
+            if (resultSet.next()) {
+                return String.format("%s:%s", resultSet.getString("Master_Host"), resultSet.getString("Master_Port"));
+            }
+            return "";
+        }
+    }
+    
     private long getSecondsBehindMaster(final Statement statement) throws SQLException {
         try (ResultSet resultSet = statement.executeQuery(SHOW_SLAVE_STATUS)) {
             if (resultSet.next()) {
@@ -145,11 +99,6 @@ public final class ShowSlaveStatusDatabaseDiscoveryType implements DatabaseDisco
             }
             return 0L;
         }
-    }
-    
-    @Override
-    public String getPrimaryDataSource() {
-        return oldPrimaryDataSource;
     }
     
     @Override
